pyproject.toml
[tool.poetry]
name = "qqbot-linger"
version = "0.0.4"
description = "My QQ Bot for personal use"
authors = ["7sDream <i@7sdre.am>"]
license = "Unlicense"
readme = "README.md"
homepage = "https://github.com/7sDream/qqbot-linger"
repository = "https://github.com/7sDream/qqbot-linger"
keywords = ["nonebot", "bilibili", "qqbot"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3.9",
    "License :: OSI Approved :: The Unlicense (Unlicense)",
]
include = [
    "UNLICENSE",
]

[tool.poetry.dependencies]
python = "^3.9"
nonebot2 = {version = "^2.0.0a10", allow-prereleases = true}
nonebot-plugin-bam = "^0.0.5"
nonebot-plugin-7s-roll = "^0.1.0"

[tool.poetry.dev-dependencies]
rope = "^0.18.0"
black = "^20.8b1"
pylint = "^2.6.0"

[tool.poetry.scripts]
linger = 'qqbot_linger.main:main'

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
